Use domain name for tunnel connections
The `borealis-pg:run` and `borealis-pg:tunnel` commands now use pg-tunnel.borealis-data.com rather than localhost for tunnel connections. The DNS A record for pg-tunnel.borealis-data.com simply points to 127.0.0.1, so the behaviour is functionally identical, but it should help to alleviate some confusion about what fields to provide to utilities like psql when connecting to the DB. For example, if one does not specify a host when running psql on a macOS/Linux machine, psql will attempt to use IPC to connect to the Postgres server because it assumes the server is installed locally. But now that `borealis-pg:tunnel` outputs a real domain name instead of "localhost", users will hopefully be compelled to provide the host rather than leaving it off.