Upgrade Node.js used by CI/CD
The Heroku CLI uses an embedded instance of Node.js. The embedded version was recently updated from Node.js 12.21.0 to 14.19.0. This repo's CI/CD workflows now use the current version of Node.js for the sake of consistency. However, the version of the `@types/node` dev dependency in `package.json` is intentionally left unchanged to ensure type compatibility for users that still have an older version of the Heroku CLI.