Support provisioning add-ons with info command
It is reasonable for a user to request details for an add-on while the add-on is still being provisioned, so the API endpoint now responds with 200 OK in that case, while simply omitting (or using a default value for) the details it does not have access to yet. The `borealis-pg:info` command has been updated to support that case as well.