Load .env file regardless of working directory
The Heroku CLI allows [linking](https://devcenter.heroku.com/articles/developing-cli-plugins#installing-the-plugin) of CLI plugins for testing/development of unpublished versions. When the plugin is linked, and one exercises the plugin using `heroku borealis-pg:...`, the process's current working directory will not be that of the plugin's repo but rather the Heroku CLI. Thus, to ensure that the `.env` file is always loaded, `dotenv` is now configured to use a path relative to the source file from which it is loaded.