Support long-running DB restores
In cases where DB restoration takes a long time (e.g. several hours), the add-on service will signal to Heroku that the add-on has finished provisioning (to avoid Heroku's 12-hour timeout for add-on provisioning: https://devcenter.heroku.com/articles/add-on-partner-api-reference#add-on-action-create-provision); meanwhile, the add-on service will continue the restore process for as long as it takes. To support this scenario, the CLI now waits for both the Heroku API and the Borealis API to report that the add-on in question is finished provisioning.