Ditch npm publish action
The `JS-DevTools/npm-publish` action was silently failing to build the project for some reason, so it has been abandoned entirely in favour of running `npm publish` directly. Also now includes [provenance](https://github.blog/security/supply-chain-security/introducing-npm-package-provenance/) for the package and allows the tag to be overridden when publishing to npm's registry.